Окно редактирования и удаления задач с подтверждением

+++ webapp/views.py
@@ -31,6 +31,15 @@ def create_task_view(request):
 def delete_task_view(request, task_id):
     task = get_object_or_404(Task, id=task_id)
 
+    if request.method == 'GET':
+        return render(
+            request,
+            'delete_task.html',
+            {
+                'task': task
+            }
+        )
+
     task.delete()
 
     return redirect('index')
@@ -43,3 +52,26 @@ def task_detail_view(request, task_id):
     }
 
     return render(request, 'task_detail.html', context)
+
+def edit_task_view(request, task_id):
+    task = get_object_or_404(Task, id=task_id)
+
+    if request.method == 'GET':
+        return render(
+            request,
+            'edit_task.html',
+            {
+                'task': task
+            }
+        )
+
+    task.description = request.POST.get('description')
+    task.details = request.POST.get('details')
+    task.status = request.POST.get('status')
+
+    due_date = request.POST.get('due_date')
+    task.due_date = due_date if due_date else None
+
+    task.save()
+
+    return redirect('task_detail', task_id=task.id)
